Revert extraction of JavaScript
